Fix: Perbaiki Owner Salary Breakdown, Tabungan Owner, dan Laba Bersih pada laporan bulanan
